+++ src/App.vue
@@ -4,6 +4,7 @@
       <router-link to="/">Home</router-link> |
       <router-link to="/about">About</router-link>
     </div>
+    <p>reserved page</p>
     <router-view/>
   </div>
 </template>