+++ src/popup/App.vue
@@ -3,16 +3,16 @@
     <h2>NotionX</h2>
     <p>A Browser Extension that enhances Notion experience.</p>
     <div>
-      <a target="_blank" href="https://github.com/scarsu/NotionX">Homepage</a>;
-      <a target="_blank" href="https://github.com/scarsu/NotionX/issues/new">Feedback Here</a>;
-      Open Source by
-      <a target="_blank" href="https://github.com/scarsu/NotionX">ScarSu</a>;
+      <li><a target="_blank" href="http://www.scarsu.cn/notionx">NotionX Homepage</a>;</li>
+      <li><a target="_blank" href="https://github.com/scarsu/NotionX/issues/new">Feedback Here</a>;</li>
+      <li>Open Source by<a target="_blank" href="https://github.com/scarsu/NotionX">ScarSu</a>;</li>
     </div>
   </div>
 </template>
 
 <script>
 
+import '@/assets/css/font.css'
 export default {
   name: 'PopUp',
   components: {},